Fix mac build 2

+++ src/cpp/buffers.cpp
@@ -124,7 +124,7 @@ DBG_EXPORT JS_METHOD(getBufferSubData) { NAPI_ENV;
 	LET_OFFS_ARG(4, length);
 	
 	size_t elementSize = dest.ElementSize();
-	size_t size = std::min(dest.ByteLength(), elementSize * length);
+	size_t size = std::min(dest.ByteLength(), static_cast<size_t>(elementSize * length));
 	size_t offset = destByteOffset * elementSize;
 	
 	void* data = getArrayData<uint8_t>(env, dest) + offset;
